feat(analytics): give account owners the traffic explorer
The admin traffic explorer becomes a shared BreakdownExplorer: the admin
page keeps every dimension, while an account-scoped tab pins its account
via hidden scope filters and drops "account" from the filter/group-by
chips so the query string can't widen the view.

Reached through a PROFILE | ANALYTICS tab on the account profile, mirroring
the product tab — the same ?tab=analytics middleware rewrite, now matching
one-segment account roots too. Carries the product page's preview callout.
