+++ src/components/features/analytics/index.ts
@@ -2,6 +2,6 @@ export { UsageCard } from "./UsageCard";
 export { UsageCardSkeleton } from "./UsageCardSkeleton";
 export { UsagePanel } from "./UsagePanel";
 export { ProductAnalyticsView } from "./ProductAnalyticsView";
-export { ProductTabs } from "./ProductTabs";
+export { ProductTabs, AccountTabs } from "./Tabs";
 export { AdminBreakdownChart } from "./AdminBreakdownChart";
 export { SERIES_COLORS, OTHER_COLOR, seriesColor } from "./palette";
