+++ src/middleware.test.ts
@@ -3,16 +3,22 @@
  * failure mode is a silently dead ANALYTICS tab — pin its behavior.
  */
 import { NextRequest } from "next/server";
-import { handleProductAnalyticsTab } from "./middleware";
+import { handleAnalyticsTab } from "./middleware";
 
 const rewriteTarget = (url: string): string | null =>
-  handleProductAnalyticsTab(new NextRequest(url))?.headers.get(
+  handleAnalyticsTab(new NextRequest(url))?.headers.get(
     "x-middleware-rewrite",
   ) ?? null;
 
 it("rewrites the product analytics tab URL to the internal route", () => {
   expect(rewriteTarget("https://source.coop/acct/prod?tab=analytics")).toBe(
     "https://source.coop/acct/prod/-/analytics",
+  );
+});
+
+it("rewrites the account analytics tab URL to the internal route", () => {
+  expect(rewriteTarget("https://source.coop/acct?tab=analytics")).toBe(
+    "https://source.coop/acct/-/analytics",
   );
 });
 
@@ -26,16 +32,16 @@ it("ignores non-matching requests", () => {
   // No tab param / wrong value
   expect(rewriteTarget("https://source.coop/acct/prod")).toBeNull();
   expect(rewriteTarget("https://source.coop/acct/prod?tab=other")).toBeNull();
-  // Not a two-segment product path
-  expect(rewriteTarget("https://source.coop/acct?tab=analytics")).toBeNull();
+  // Deeper than an account or product root
   expect(
     rewriteTarget("https://source.coop/acct/prod/file.txt?tab=analytics"),
   ).toBeNull();
-  // Two-segment top-level app routes are not products
+  // Top-level app routes are neither accounts nor products
   expect(
     rewriteTarget("https://source.coop/admin/analytics?tab=analytics"),
   ).toBeNull();
   expect(
     rewriteTarget("https://source.coop/products/new?tab=analytics"),
   ).toBeNull();
+  expect(rewriteTarget("https://source.coop/products?tab=analytics")).toBeNull();
 });
